refactor indexing console command
