+++ app/Console/Commands/EsIndexDocuments.php
@@ -25,7 +25,7 @@ use App\Models\Beatmapset;
 use App\Models\Forum\Post;
 use Illuminate\Console\Command;
 
-class EsIndexDocuments extends Command
+class EsIndexDocuments extends EsIndexCommand
 {
     /**
      * The name and signature of the console command.
@@ -41,111 +41,6 @@ class EsIndexDocuments extends Command
      */
     protected $description = 'Indexes documents into Elasticsearch.';
 
-    private $cleanup;
-    private $inplace;
-    private $types;
-    private $suffix;
-
-    /**
-     * Create a new command instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
-    /**
-     * Execute the console command.
-     *
-     * @return mixed
-     */
-    public function handle()
-    {
-        $this->readOptions();
-        $this->types = [Beatmapset::class, Post::class];
-        $this->suffix = !$this->inplace ? '_'.time() : '';
-
-        $oldIndices = Indexing::getOldIndices('osu');
-
-        $continue = $this->starterMessage($oldIndices);
-        if (!$continue) {
-            return $this->error('User aborted!');
-        }
-
-        $indices = $this->index();
-
-        $this->finish($indices, $oldIndices);
-        $this->warn("\nIndexing completed.");
-    }
-
-    private function finish(array $indices, array $oldIndices)
-    {
-        // always update osu alias
-        $indicesString = implode(', ', $indices);
-        $this->warn("Aliasing '{$indicesString}' to 'osu'...");
-        Indexing::updateAlias('osu', $indices);
-
-        if (!$this->inplace && $this->cleanup) {
-            foreach ($oldIndices as $index) {
-                $this->warn("Removing '{$index}'...");
-                Indexing::deleteIndex($index);
-            }
-        }
-    }
-
-    /**
-     * Indexes and returns the names of the indices.
-     *
-     * @return array names of the indices indexed to.
-     */
-    private function index()
-    {
-        $indices = [];
-        foreach ($this->types as $type) {
-            if (!$this->inplace) {
-                $indexName = "{$type::esIndexName()}{$this->suffix}";
-
-                $this->info("Indexing {$type} into {$indexName}");
-                $type::esIndexIntoNew(1000, $indexName);
-
-                $indices[] = $indexName;
-            } else {
-                $this->info("In-place indexing {$type} into {$type::esIndexName()}");
-                $type::esReindexAll(1000);
-
-                $indices[] = $type::esIndexName();
-            }
-        }
-
-        return $indices;
-    }
-
-    private function readOptions()
-    {
-        $this->inplace = $this->option('inplace');
-        $this->cleanup = $this->option('cleanup');
-    }
-
-    private function starterMessage(array $oldIndices)
-    {
-        if ($this->inplace) {
-            $this->warn('Running in-place reindex.');
-            $confirmMessage = "This will reindex in-place (schemas must match) and alias them to 'osu'";
-        } else {
-            $this->warn('Running index transfer.');
-
-            if ($this->cleanup) {
-                $this->warn(
-                    "The following indices will be deleted on completion!\n"
-                    .implode("\n", $oldIndices)
-                );
-            }
-
-            $confirmMessage = "This will create new indices and alias them to 'osu'";
-        }
-
-        return $this->confirm("{$confirmMessage}, begin indexing?");
-    }
+    protected $alias = 'osu';
+    protected $types = [Beatmapset::class, Post::class];
 }
